+++ types/window.ts
@@ -1,7 +1,7 @@
 import { URI } from 'vscode-languageserver-types'
 
 export const SHOW_SAVE_FILE_DIALOG_REQUEST_METHOD = 'aws/showSaveFileDialog'
-
+export const SHOW_OPEN_FILE_DIALOG_REQUEST_METHOD = 'aws/showOpenFileDialog'
 export interface ShowSaveFileDialogParams {
     // Using untyped string to avoid locking this too strictly.
     // TODO: Migrate to LanguageKind when it is released in 3.18.0
@@ -12,4 +12,18 @@ export interface ShowSaveFileDialogParams {
 
 export interface ShowSaveFileDialogResult {
     targetUri: URI
+}
+
+// bridge to consume ide's api
+export interface ShowOpenDialogParams {
+    canSelectFiles?: boolean
+    canSelectFolders?: boolean
+    canSelectMany?: boolean
+    filters?: { [key: string]: string[] }
+    defaultUri?: URI
+    title?: string
+}
+
+export interface ShowOpenDialogResult {
+    uris: URI[]
 }